Refactor of build.py and build/deploy process
Signed Rylogic assemblies
Updatednuget packages
netcoreapp3.1 support
renamed Range to RangeI to avoid collision with System.Range

+++ projects/KeySpy/src/main.cpp
@@ -87,7 +87,7 @@ namespace keyspy
 			}
 
 			// Write 1 bit to record the Alt key state
-			PutBits(AllSet(p.flags, LLKHF_ALTDOWN) ? 1 : 0, 1);
+			PutBits(AllSet(p.flags, DWORD(LLKHF_ALTDOWN)) ? 1 : 0, 1);
 
 			// Write 8 bits for the virtual key code
 			// 'vkCode - 1' = a value on the range [0,253]
